Refactor game state handling and network communication
- Removed the HandleShips class and integrated its functionality into the new HandleShips handler.
- Introduced new handlers for managing player actions, cell reveals, ship placements, and resource updates.
- Updated ClientService to utilize the new handlers for processing server messages.
- Simplified the GameStateManager by removing unused namespaces and adding XML documentation for clarity.
- Enhanced the GameBoard rendering logic to improve visual feedback for hits and misses.
- Consolidated lobby handling logic into a dedicated LobbyHandler class for better organization.
- Removed obsolete code and files to streamline the project structure.
