+++ src/WaterWizard.Client/gamescreen/handler/HandleShips.cs
@@ -1,16 +1,19 @@
 using LiteNetLib;
-using WaterWizard.Client.gamescreen;
+using WaterWizard.Client.gamescreen.ships;
 
-namespace WaterWizard.Client.gamescreen.ships;
+namespace WaterWizard.Client.gamescreen.handler;
 
+/// <summary>
+/// Handles ship-related messages received from the server.
+/// </summary>
 public class HandleShips
 {
-   /// <summary>
+    /// <summary>
     /// Handles the ship position message received from the server.
     /// </summary>
     /// <param name="messageType">Message Type that is getting received from the Server</param>
     /// <param name="reader">The NetPacketReader containing the serialized ship data sent from the server</param>
-public static void HandleShipPosition(string messageType, NetPacketReader reader)
+    public static void HandleShipPosition(string messageType, NetPacketReader reader)
     {
         var playerBoard = GameStateManager.Instance.GameScreen!.playerBoard;
         if (playerBoard == null)
@@ -82,5 +85,64 @@ public static void HandleShipPosition(string messageType, NetPacketReader reader
         Console.WriteLine($"[Client] Nach ShipSync sind {playerBoard.Ships.Count} Schiffe auf dem Board.");
         GameStateManager.Instance.SetStateToInGame();
         Console.WriteLine($"[Client] Nach SetStateToInGame sind {playerBoard.Ships.Count} Schiffe auf dem Board.");
-    } 
+    }
+
+    /// <summary>
+    /// Handles the ship reveal message received from the server.
+    /// </summary>
+    /// <param name="reader">The NetPacketReader containing the serialized ship data sent from the server</param>
+    public static void HandleShipReveal(NetPacketReader reader)
+    {
+        int x = reader.GetInt();
+        int y = reader.GetInt();
+        int width = reader.GetInt();
+        int height = reader.GetInt();
+
+        var opponentBoard = GameStateManager.Instance.GameScreen!.opponentBoard;
+        if (opponentBoard != null)
+        {
+            int pixelX = (int)opponentBoard.Position.X + x * opponentBoard.CellSize;
+            int pixelY = (int)opponentBoard.Position.Y + y * opponentBoard.CellSize;
+            int pixelWidth = width * opponentBoard.CellSize;
+            int pixelHeight = height * opponentBoard.CellSize;
+
+            opponentBoard.putShip(
+                new GameShip(
+                    GameStateManager.Instance.GameScreen,
+                    pixelX,
+                    pixelY,
+                    ShipType.DEFAULT,
+                    pixelWidth,
+                    pixelHeight
+                )
+            );
+
+            Console.WriteLine(
+                $"[Client] Ship revealed: ({x},{y}) size {width}x{height}"
+            );
+        }
+    }
+
+    /// <summary>
+    /// Handles the ship placement error message received from the server.
+    /// </summary>
+    /// <param name="reader">The NetPacketReader containing the serialized ship data sent from the server</param>
+    public static void HandleShipPlacementError(NetPacketReader reader)
+    {
+        string errorMsg = reader.GetString();
+        Console.WriteLine(
+            $"[Client] Fehler beim Platzieren des Schiffs: {errorMsg}"
+        );
+
+        // Sperre das Draggen für diese Größe, wenn das Limit erreicht ist
+        var match = System.Text.RegularExpressions.Regex.Match(
+            errorMsg,
+            @"nur (\d+) Schiffe der Länge (\d+)"
+        );
+        if (match.Success)
+        {
+            int size = int.Parse(match.Groups[2].Value);
+            GameStateManager.Instance.GameScreen?.MarkShipSizeLimitReached(size);
+        }
+    }
 }
